+++ relm4-macros/src/component/mod.rs
@@ -91,11 +91,11 @@ pub(crate) fn generate_tokens(
             #rename_root
             #menus_stream
             #init_widgets
+            #assign
             #connect
             {
                 #error
             }
-            #assign
         };
 
         let widgets_return_code = parse_quote! {