+++ relm4-macros/src/view.rs
@@ -27,8 +27,8 @@ pub(super) fn generate_tokens(input: TokenStream) -> TokenStream {
 
     let output = quote! {
         #init
-        #connect
         #assign
+        #connect
         {
             #error
         }