+++ relm4-macros/src/widgets/gen/conditional_init.rs
@@ -165,11 +165,11 @@ impl AssignProperty {
             match &self.attr {
                 AssignPropertyAttr::None => (),
                 AssignPropertyAttr::Watch => {
-                    self.assign_stream(stream, p_name, w_name);
+                    self.assign_stream(stream, p_name, w_name, true);
                 }
                 AssignPropertyAttr::Track((track_stream, paste_model)) => {
                     let mut assign_stream = TokenStream2::new();
-                    self.assign_stream(&mut assign_stream, p_name, w_name);
+                    self.assign_stream(&mut assign_stream, p_name, w_name, true);
                     let model = paste_model.then(|| model_name);
 
                     stream.extend(quote_spanned! {